+++ includes/requirements.php
@@ -1,19 +1,21 @@
 <?php
-/*
-	Copyright 2014 Michael Cannon (email: mc@aihr.us)
+/**
+Aihrus WordPress Starter Premium
+Copyright (C) 2014  Michael Cannon
 
-	This program is free software; you can redistribute it and/or modify
-	it under the terms of the GNU General Public License, version 2, as
-	published by the Free Software Foundation.
+This program is free software; you can redistribute it and/or modify
+it under the terms of the GNU General Public License as published by
+the Free Software Foundation; either version 2 of the License, or
+(at your option) any later version.
 
-	This program is distributed in the hope that it will be useful,
-	but WITHOUT ANY WARRANTY; without even the implied warranty of
-	MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-	GNU General Public License for more details.
+This program is distributed in the hope that it will be useful,
+but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+GNU General Public License for more details.
 
-	You should have received a copy of the GNU General Public License
-	along with this program; if not, write to the Free Software
-	Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA  02110-1301  USA
+You should have received a copy of the GNU General Public License along
+with this program; if not, write to the Free Software Foundation, Inc.,
+51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
  */
 
 require_once AIHR_DIR . 'aihrus-framework.php';
@@ -42,7 +44,7 @@ function wpsp_requirements_check( $force_check = false ) {
 	global $wps_activated;
 
 	if ( empty( $wps_activated ) ) {
-		$deactivate_reason = esc_html__( 'Internal WordPress Starter not detected' );
+		$deactivate_reason = esc_html__( 'Internal Aihrus WordPress Starter not detected' );
 	}
 
 	if ( ! empty( $deactivate_reason ) ) {
@@ -50,8 +52,10 @@ function wpsp_requirements_check( $force_check = false ) {
 	}
 	
 	$check_okay = empty( $deactivate_reason );
-	delete_transient( 'wpsp_requirements_check' );
-	set_transient( 'wpsp_requirements_check', $check_okay, WEEK_IN_SECONDS );
+	if ( $check_okay ) {
+		delete_transient( 'wpsp_requirements_check' );
+		set_transient( 'wpsp_requirements_check', $check_okay, HOUR_IN_SECONDS );
+	}
 
 	return $check_okay;
 }